Replace c_int with isize throughout codebase
Eliminate all uses of c_int in favor of Rust's native isize type for
I/O operations, file descriptors, and handles. This simplifies the code
by removing unnecessary type conversions while maintaining full compatibility.

Rationale:
- On wasm32 targets (primary FFI target), c_int and isize are both i32
- isize is Rust's native signed size type, more idiomatic
- Eliminates ALL conversions between c_int and isize
- POSIX read/write semantics use signed return values (ssize_t pattern)

Changes across all crates:

actor_runtime changes:
- ActorRuntime trait: all methods use isize instead of c_int
- FFI extern declarations: changed from c_int to isize (same ABI on wasm32)
- FfiActorRuntime: no longer needs conversions, passes isize directly to FFI
- dagops: all handle and fd types changed from i32 to isize

actor_io changes:
- error_mapping: errno_to_error_kind now takes isize
- AReader/AWriter: fd fields and parameters changed to isize
- Removed unnecessary c_int imports

ailetos changes:
- stub_actor_runtime: all ActorRuntime impl methods use isize
- system_runtime: removed c_int import and usage entirely
- pipe: kept try_from conversion for usize->isize (needed for 64-bit safety)

Type flow now completely unified:
- data.len() (usize) -> Writer::write() (isize) via try_from
- Internal I/O (isize throughout, no conversions)
- FFI boundary (isize, matches wasm32 ABI as i32)

+++ ailets-rs/actor_io/src/areader.rs
@@ -24,12 +24,11 @@
 //! ```
 
 use actor_runtime::{ActorRuntime, StdHandle};
-use core::ffi::c_int;
 
 use crate::error_mapping::errno_to_error_kind;
 
 pub struct AReader<'a> {
-    fd: Option<c_int>,
+    fd: Option<isize>,
     runtime: &'a dyn ActorRuntime,
 }
 
@@ -58,7 +57,7 @@ impl<'a> AReader<'a> {
     #[must_use]
     pub fn new_from_std(runtime: &'a dyn ActorRuntime, handle: StdHandle) -> Self {
         Self {
-            fd: Some(handle as c_int),
+            fd: Some(handle as isize),
             runtime,
         }
     }
